pyproject.toml
[project]
name = "gitscout"
version = "0.1.0"
description = "Uses AI to help understand repositories and their changes."
readme = "README.md"
authors = [{name = "James Stout"}]
requires-python = ">=3.11"
classifiers = [
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "click",
    "flask>=3.1.0",
    "gitpython>=3.1.43",
    "llm-gemini>=0.4.2",
    "llm>=0.18",
    "mistune>=3.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/wolfmanstout/gitscout"
Changelog = "https://github.com/wolfmanstout/gitscout/releases"
Issues = "https://github.com/wolfmanstout/gitscout/issues"
CI = "https://github.com/wolfmanstout/gitscout/actions"

[project.scripts]
gitscout = "gitscout.cli:cli"

[dependency-groups]
dev = [
    "pytest",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["E501", "SIM105", "SIM116", "UP007"]
